+++ errors.hpp
@@ -19,3 +19,5 @@
 #define EXIT_UNLINK 118
 #define EXIT_UTF8 119
 #define EXIT_WRITE 120
+
+// avoid 124, 125, 126, 127, 137, which are used by GNU timeout