+++ options.hpp
@@ -26,6 +26,7 @@
 #define A_CONVERT_NUMERIC_IDS ((int) 'I')
 #define A_HILBERT ((int) 'h')
 #define A_VISVALINGAM ((int) 'v')
+#define A_GENERATE_POLYGON_LABEL_POINTS ((int) 'P')
 
 #define P_SIMPLIFY ((int) 's')
 #define P_SIMPLIFY_LOW ((int) 'S')