+++ serial.hpp
@@ -56,6 +56,7 @@ struct serial_feature {
 
 	drawvec geometry = drawvec();
 	unsigned long long index = 0;
+	unsigned long long label_point = 0;
 	long long extent = 0;
 
 	std::vector<long long> keys{};